Added gold display and inventory guide

+++ zuul-framework/src/TWoT_A1/TWoT.java
@@ -921,7 +921,7 @@ public class TWoT implements Serializable {
      *
      * @return
      */
-    public double getPlayerHealth() {
+    public int getPlayerHealth() {
         return player.getHealth();
     }
 
@@ -929,7 +929,7 @@ public class TWoT implements Serializable {
      *
      * @return
      */
-    public double getPlayerGold() {
+    public int getPlayerGold() {
         return player.getGold();
     }
